requirements.txt
# from gphoto2 install

backcall==0.1.0
decorator==4.4.0
gphoto2==2.0.0
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7

# stepper controller install

Adafruit-Blinka==2.2.0
adafruit-circuitpython-busdevice==3.0.0
adafruit-circuitpython-motor==2.0.0
adafruit-circuitpython-motorkit==1.3.0
adafruit-circuitpython-pca9685==3.2.5
adafruit-circuitpython-register==1.7.1
Adafruit-PlatformDetect==1.2.1
Adafruit-PureIO==0.2.3

rpi-ws281x==4.2.2
RPi.GPIO==0.7.0

spidev==3.4
sysv-ipc==1.0.0
python-dotenv
boto3
